WIP Changes for S3 data writes.
