+++ service/Cargo.toml
@@ -8,7 +8,7 @@ edition = "2018"
 
 [dependencies]
 jemallocator = "0.3.2"
-anyhow = "1.0.42"
+anyhow = "1.0.43"
 serde = { version = "1.0.127", features = ['derive', 'rc'] }
 serde_json = "1.0.66"
 log = "0.4.14"
@@ -19,9 +19,9 @@ rand = "0.8.4"
 config = "0.11.0"
 base64 = "0.13.0"
 clap = { version = "2.33.3", features = ["yaml"] }
-log4rs = {version = "1.0.0", features = ["background_rotation"] }
+log4rs = { version = "1.0.0", features = ["background_rotation"] }
 hyper = { version = "0.14.11", features = ["full"] }
-tokio = { version = "1.9.0", features = ["full"] }
+tokio = { version = "1.10.0", features = ["full"] }
 itertools = "0.10.1"
 diesel = { version = "1.4.7", features = ["extras", "postgres", "uuidv07"] }
 enumset = "1.0.7"
@@ -58,6 +58,7 @@ regex = "1.5.4"
 fasthash = "0.4.0"
 
 validator = { version = "0.14.0", features = ["derive"] }
+rust-s3 = "0.26.4"
 
 [dependencies.chrono]
 version = "0.4.19"
