+++ service/examples/sample.json
@@ -1,8 +1,8 @@
 [
   {
-    "id": "IRenRc-iFWegvEqGAPrJy",
-    "name": "NoScript-relaunch",
-    "short_name": "NS-rl",
+    "id": "9xriNtdYnaH0AK1GV1Cju",
+    "name": "onboarding",
+    "short_name": "onb",
     "version": 1,
     "kind": "Experiment",
     "inactive": false,
@@ -10,14 +10,14 @@
     "end_time": null,
     "audiences": [
       {
-        "name": "nsc_rlnch_users",
-        "script_src": null,
+        "name": "new_users",
+        "script_src": "ctx.get('new_user', False) and sem_version(ctx.get(app_version, '0.0.0')) >= '4.7.3'",
         "list_id": null,
         "size_kind": "Percent",
         "value": 100
       }
     ],
-    "frequency_constraint": null,
+    "frequency_constraint": "experiment.get(total_invocation_count, 0) % 2 == 0",
     "variations": null,
     "data": {
       "flow_config": [
@@ -32,9 +32,9 @@
     }
   },
   {
-    "id": "js_8qLsqGaixv2aPaAtsR",
-    "name": "Noscript-Onboarding",
-    "short_name": "ns-on",
+    "id": "kM_KtKLGmMe4U4oK5reay",
+    "name": "relaunch",
+    "short_name": "rlnch",
     "version": 2,
     "kind": "Experiment",
     "inactive": false,
@@ -42,25 +42,56 @@
     "end_time": null,
     "audiences": [
       {
-        "name": "nsc_new_users",
-        "script_src": null,
+        "name": "new_users",
+        "script_src": "not ctx.get('new_user', False) and sem_version(ctx.get(app_version, '0.0.0')) >= '4.7.3'",
         "list_id": null,
         "size_kind": "Percent",
+        "value": 100
+      }
+    ],
+    "frequency_constraint": "experiment.get(total_invocation_count, 0) % 2 == 0",
+    "variations": null,
+    "data": {
+      "flow_config": [
+        "l_s",
+        "s_i",
+        "p_b_i",
+        "p_p_i",
+        "i_s",
+        "c_s",
+        "f"
+      ]
+    }
+  },
+  {
+    "id": "tWJvvVeKwu6-ia2Z4_Q2-",
+    "name": "cronet",
+    "short_name": "crnt",
+    "version": 7,
+    "kind": "Experiment",
+    "inactive": false,
+    "start_time": null,
+    "end_time": null,
+    "audiences": [
+      {
+        "name": "new_users",
+        "script_src": "sem_version(ctx.get(app_version, '0.0.0')) >= '4.4.0'",
+        "list_id": null,
+        "size_kind": "Percent",
+        "value": 10
+      },
+      {
+        "name": "whitelist",
+        "script_src": null,
+        "list_id": "nWDppTrcmJxlx6kMXwHS5",
+        "size_kind": "Absolute",
         "value": 100
       }
     ],
     "frequency_constraint": null,
     "variations": null,
     "data": {
-      "flow_config": [
-        "s_i",
-        "l_s",
-        "c_s",
-        "p_b_i",
-        "p_p_i",
-        "i_s",
-        "f"
-      ]
+      "da_factory": "cronet"
     }
   }
 ]
